code/core_product/requirements.txt
Flask==1.0.2
numpy==1.15.4
pandas==0.24.1
pyenchant==2.0.0
langdetect==1.0.7
scikit-learn==0.20.2
google-api-python-client==1.7.8
google-cloud-storage==1.13.2
spacy==2.0.13
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.0.0/en_core_web_sm-2.0.0.tar.gz#en_core_web_sm#egg=en_core_web_sm==2.0.0